Reorganized project structure and refactored import statements
Several files were moved to different folders to better reflect their purpose and function in the project. The import statements in these files were updated to match the new project structure. Unnecessary comments were also removed from SMSService.java.
