+++ backend/src/main/java/com/example/backend/account/transfer/TransferRequest.java
@@ -1,5 +1,6 @@
-package com.example.backend.entity;
+package com.example.backend.account.transfer;
 
+import com.example.backend.account.Account;
 import com.fasterxml.jackson.annotation.JsonBackReference;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
