+++ frontend/src/App.tsx
@@ -1,14 +1,14 @@
 import {Routes, Route} from "react-router-dom";
 import './App.css'
 
-import LoginPage from "./pages/LoginPage.tsx";
-import RegisterPage from "./pages/RegisterPage.tsx";
-import Home from "./pages/Home.tsx";
+import LoginPage from "./pages/authPages/LoginPage.tsx";
+import RegisterPage from "./pages/authPages/RegisterPage.tsx";
+import Home from "./pages/userPages/Home.tsx";
 import WelcomePage from "./pages/WelcomePage.tsx";
-import NotAuthorised from "./pages/Not-Authorised.tsx";
-import TransferForm from "./pages/TransferForm.tsx";
-import AccountTransactions from "./pages/AccountTransactions.tsx";
-import CurrencyConverterForm from "./pages/CurrencyConverterForm.tsx";
+import NotAuthorised from "./pages/authPages/Not-Authorised.tsx";
+import TransferForm from "./pages/userPages/TransferForm.tsx";
+import AccountTransactions from "./pages/userPages/AccountTransactions.tsx";
+import CurrencyConverterForm from "./pages/userPages/CurrencyConverterForm.tsx";
 
 
 function App() {
