+++ frontend/src/components/navigationBars/HomeNavBar.tsx
@@ -1,8 +1,8 @@
 import { Link, useNavigate } from "react-router-dom";
 import { motion } from "framer-motion";
-import BankingLogo from "../assets/BankingLogo.png";
+import BankingLogo from "../../assets/BankingLogo.png";
 
-export default function NavBar() {
+export default function HomeNavBar() {
     const navigate = useNavigate();
 
     const handleLogout = () => {
