+++ frontend/src/components/security/ProtectedRoute.tsx
